fixed: Some compile errors.

+++ ascension/ascension/graphics/physical-direction.hpp
@@ -40,7 +40,7 @@ namespace ascension {
 				PhysicalDirection::BOTTOM, PhysicalDirection::LEFT,
 				PhysicalDirection::TOP, PhysicalDirection::RIGHT
 			};
-			const std::size_t index = direction;
+			const std::size_t index = static_cast<std::size_t>(direction);
 			if(index >= std::extent<decltype(opposites)>::value)
 				throw UnknownValueException("direction");
 			return opposites[index];
